% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documention.r
\docType{data}
\name{chemi_aa_coldf}
\alias{chemi_aa_coldf}
\title{color scheme of sequence which are amino acids aligned according to chemistry}
\format{A data frame with 6 rows and 3 variables}
\usage{
data(chemi_aa_coldf)
}
\description{
A dataset containing the color scheme of 6 observations of 3
variables. The variables are as follows:
}
\details{
\itemize{
   \item{\code{amino_acid}}{a character vector}
   \item{\code{color}}{a character vector}
   \item{\code{chemistry}}{a character vector}
}
}
\keyword{datasets}
