% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documention.r
\docType{data}
\name{chemi_nucle_coldf}
\alias{chemi_nucle_coldf}
\title{color scheme of sequence which are nucleotides aligned according to chemistry}
\format{A data frame with 5 rows and 2 variables}
\usage{
data(chemi_nucle_coldf)
}
\description{
A dataset containing the color scheme of 5 observations of 2
variables. The variables are as follows:
}
\details{
\itemize{
   \item{\code{nucleotide}}{a character vector}
   \item{\code{color}}{a character vector}
}
}
\keyword{datasets}
