% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorscheme.R
\name{color_scheme}
\alias{color_scheme}
\title{colorscheme}
\usage{
color_scheme(seq, scheme_df)
}
\arguments{
\item{seq}{aligned data frame}

\item{scheme_df}{color scheme data frame}
}
\value{
color matched with sequence
}
\description{
colorscheme
}
\examples{
color_scheme(sampleseq,col_df)
}
\author{
huina huang
}
