% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_colorscheme.R
\name{msa_color_sheme}
\alias{msa_color_sheme}
\title{msa_color_sheme}
\usage{
msa_color_sheme(seq, scheme_df)
}
\arguments{
\item{seq}{aligned data frame}

\item{scheme_df}{color scheme data frame}
}
\value{
color matched with sequence
}
\description{
msa_color_sheme
}
\examples{
msa_color_sheme(sampleseq, "chemi_aa_coldf")
}
\author{
huina huang
}
