% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documention.r
\docType{data}
\name{sampleseqcolor}
\alias{sampleseqcolor}
\title{color of sampleseq according to col_df}
\format{A data frame with 9 rows and 50 variables}
\usage{
data(sampleseqcolor)
}
\description{
A dataset containing the test sequence of 9 observations of 50
variables. The variables are as follows:
}
\details{
\itemize{
   \item{\code{V164}}{a character vector}
   \item{\code{V165}}{a character vector}
   \item{\code{V166}}{a character vector}
   \item{\code{V167}}{a character vector}
   \item{\code{V168}}{a character vector}
   \item{\code{V169}}{a character vector}
   \item{\code{V170}}{a character vector}
   \item{\code{V171}}{a character vector}
   \item{\code{V172}}{a character vector}
   \item{\code{V173}}{a character vector}
   \item{\code{V174}}{a character vector}
   \item{\code{V175}}{a character vector}
   \item{\code{V176}}{a character vector}
   \item{\code{V177}}{a character vector}
   \item{\code{V178}}{a character vector}
   \item{\code{V179}}{a character vector}
   \item{\code{V180}}{a character vector}
   \item{\code{V181}}{a character vector}
   \item{\code{V182}}{a character vector}
   \item{\code{V183}}{a character vector}
   \item{\code{V184}}{a character vector}
   \item{\code{V185}}{a character vector}
   \item{\code{V186}}{a character vector}
   \item{\code{V187}}{a character vector}
   \item{\code{V188}}{a character vector}
   \item{\code{V189}}{a character vector}
   \item{\code{V190}}{a character vector}
   \item{\code{V191}}{a character vector}
   \item{\code{V192}}{a character vector}
   \item{\code{V193}}{a character vector}
   \item{\code{V194}}{a character vector}
   \item{\code{V195}}{a character vector}
   \item{\code{V196}}{a character vector}
   \item{\code{V197}}{a character vector}
   \item{\code{V198}}{a character vector}
   \item{\code{V199}}{a character vector}
   \item{\code{V200}}{a character vector}
   \item{\code{V201}}{a character vector}
   \item{\code{V202}}{a character vector}
   \item{\code{V203}}{a character vector}
   \item{\code{V204}}{a character vector}
   \item{\code{V205}}{a character vector}
   \item{\code{V206}}{a character vector}
   \item{\code{V207}}{a character vector}
   \item{\code{V208}}{a character vector}
   \item{\code{V209}}{a character vector}
   \item{\code{V210}}{a character vector}
   \item{\code{V211}}{a character vector}
   \item{\code{V212}}{a character vector}
   \item{\code{V213}}{a character vector}
}
}
\keyword{datasets}
