% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documention.r
\docType{data}
\name{shape_nucle_coldf}
\alias{shape_nucle_coldf}
\title{color scheme of sequence which are nucleotides aligned according to the RasMol 'shapely models'}
\format{A data frame with 6 rows and 2 variables}
\usage{
data(shape_nucle_coldf)
}
\description{
A dataset containing the color scheme of 6 observations of 2
variables. The variables are as follows:
}
\details{
\itemize{
   \item{\code{nucleotide}}{a character vector}
   \item{\code{color}}{a character vector}
}
}
\keyword{datasets}
