% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documention.r
\docType{data}
\name{testseqcolor_shape}
\alias{testseqcolor_shape}
\title{color of testseq according to shape_aa_coldf}
\format{A data frame with 9 rows and 7 variables}
\usage{
data(testseqcolor_shape)
}
\description{
A dataset containing the test sequence of 9 observations of 7
variables. The variables are as follows:
}
\details{
\itemize{
   \item{\code{V1}}{a character vector}
   \item{\code{V2}}{a character vector}
   \item{\code{V3}}{a character vector}
   \item{\code{V4}}{a character vector}
   \item{\code{V5}}{a character vector}
   \item{\code{V6}}{a character vector}
   \item{\code{V7}}{a character vector}
}
}
\keyword{datasets}
