% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documention.r
\docType{data}
\name{Zappo_aa_coldf}
\alias{Zappo_aa_coldf}
\title{color scheme of sequence which are amino acids aligned according to Zappo}
\format{A data frame with 8 rows and 3 variables}
\usage{
data(Zappo_aa_coldf)
}
\description{
A dataset containing the color scheme of 8 observations of 2
variables. The variables are as follows:
}
\details{
\itemize{
   \item{\code{nucleotide}}{a character vector}
   \item{\code{color}}{a character vector}
   \item{\code{physicochemical_properties}}{a character vector}
}
}
\keyword{datasets}
